requirements.txt
google-generativeai==0.5.0
PyGithub==2.3.0
httpx==0.28.1
